Refined code to support AWS lambda.
