+++ Makefile
@@ -1,6 +1,6 @@
-TARGETS := agraffe 
+TARGETS := agraffe
 
-ISORT_OPT := -m3 -tc --py 37
+ISORT_OPT := -m3 --tc --py 37
 BLACK_OPT := -t py37 --skip-string-normalization
 FLAKE8_OPT := --max-line-length 88
 MYPY_OPT := --config-file mypy.ini
