+++ example/requirements.txt
@@ -1,0 +1,1 @@
+fastapi==0.61.0
